refactor: repository 디렉토리명 변경 및 변경적용
